Move and consolidate label-filtering tests into test_relational_query_match_sdata_to_table
- Replace test_match_sdata_to_table_match_labels_error with
  test_filter_out_instances: parametrized over both API functions and
  element types; tests all three filter_label_pixels states (None→warn,
  False→nullcontext noop, True→pixels filtered)
- Add test_subset_sdata_by_table_mask for mixed-element subsetting
- Delete test_relational_query_subset_sdata_by_table_mask.py
